backend/requirements.txt
absl-py==0.9.0
amqp==2.6.0
asgiref==3.2.7
astor==0.8.1
autopep8==1.5.2
billiard==3.6.3.0
celery==4.4.4
cffi==1.14.0
cryptography==2.9.2
cycler==0.10.0
Django==3.0.5
django-cors-headers==3.2.1
django-environ==0.4.5
-e git+https://github.com/confirm/django-fulltext-search.git@1ab6f3b7dcd9fb7ca9d8bc5f400184e334699e3d#egg=django_fulltext_search
django-rest-knox==4.1.0
djangorestframework==3.11.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
imageai==2.1.5
importlib-metadata==1.6.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
kombu==4.6.10
lxml==4.5.1
Markdown==3.2.2
matplotlib==3.2.1
mysqlclient==1.4.6
numpy==1.18.4
opencv-python==4.2.0.34
opt-einsum==3.2.1
Pillow==7.1.2
protobuf==3.12.2
pycodestyle==2.6.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-http-client==3.2.7
pytz==2020.1
PyYAML==5.3.1
scipy==1.4.1
sendgrid==6.3.1
six==1.14.0
sqlparse==0.3.1
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
vine==1.3.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
